+++ HISTORY.md
@@ -1,3 +1,16 @@
 History
 
 ---
+
+## 1.0.0
+
+### API
+
+1. remove `openClassName` `width`;
+2. update `iconChild` -> `handleChild` `onIconClick` -> `onHandleClick`;
+3. add `handleStyle` `showMask`;
+
+### Func
+
+1. add `top` `bottom` placement.
+2. `handleChild` type remove Array;